도서등록 ajax 진행중

+++ stsbmvc0_book/src/main/java/com/touzone/stsb/repository/BookRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.touzone.stsb.vo.BookVo;
 
-public interface BookRepository extends JpaRepository<BookVo, Long>{
+public interface BookRepository extends JpaRepository<BookVo, Integer>{
 
 }
